feat: refactor into more files
